+++ packages/entities/entities-redis-configurations/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.15.9](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-redis-configurations@1.15.8...@kong-ui-public/entities-redis-configurations@1.15.9) (2026-05-15)
+
+**Note:** Version bump only for package @kong-ui-public/entities-redis-configurations
+
+
+
+
+
+## [1.15.8](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-redis-configurations@1.15.7...@kong-ui-public/entities-redis-configurations@1.15.8) (2026-05-15)
+
+**Note:** Version bump only for package @kong-ui-public/entities-redis-configurations
+
+
+
+
+
 ## [1.15.7](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-redis-configurations@1.15.6...@kong-ui-public/entities-redis-configurations@1.15.7) (2026-05-13)
 
 **Note:** Version bump only for package @kong-ui-public/entities-redis-configurations