+++ packages/entities/entities-routes/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.22.11](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-routes@3.22.10...@kong-ui-public/entities-routes@3.22.11) (2026-05-15)
+
+**Note:** Version bump only for package @kong-ui-public/entities-routes
+
+
+
+
+
+## [3.22.10](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-routes@3.22.9...@kong-ui-public/entities-routes@3.22.10) (2026-05-15)
+
+**Note:** Version bump only for package @kong-ui-public/entities-routes
+
+
+
+
+
 ## [3.22.9](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-routes@3.22.8...@kong-ui-public/entities-routes@3.22.9) (2026-05-13)
 
 **Note:** Version bump only for package @kong-ui-public/entities-routes